project-bolt-sb1-ffe998dq1/project/backend/ml/requirements.txt
tensorflow>=2.10.0
fastapi>=0.68.0
uvicorn>=0.15.0
pillow>=8.3.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
python-multipart>=0.0.5
